requirements.txt
pynput
cursor
colorama
python-cfonts